requirements.txt
awsebcli==3.14.10
botocore==1.12.86
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
Django==2.1.5
docutils==0.14
idna==2.7
jmespath==0.9.3
pathspec==0.5.5
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
urllib3==1.24.1
django-widget-tweaks

